+++ lib/zclient.h
@@ -241,7 +241,6 @@ struct zclient {
 	void (*zebra_capabilities)(struct zclient_capabilities *cap);
 	int (*router_id_update)(ZAPI_CALLBACK_ARGS);
 	int (*interface_delete)(ZAPI_CALLBACK_ARGS);
-	int (*interface_up)(ZAPI_CALLBACK_ARGS);
 	int (*interface_down)(ZAPI_CALLBACK_ARGS);
 	int (*interface_address_add)(ZAPI_CALLBACK_ARGS);
 	int (*interface_address_delete)(ZAPI_CALLBACK_ARGS);
