+++ ripngd/ripng_zebra.c
@@ -242,7 +242,6 @@ void zebra_init(struct thread_master *master)
 	zclient_init(zclient, ZEBRA_ROUTE_RIPNG, 0, &ripngd_privs);
 
 	zclient->zebra_connected = ripng_zebra_connected;
-	zclient->interface_up = ripng_interface_up;
 	zclient->interface_down = ripng_interface_down;
 	zclient->interface_delete = ripng_interface_delete;
 	zclient->interface_address_add = ripng_interface_address_add;
