+++ sharpd/sharp_zebra.c
@@ -46,17 +46,6 @@ struct zclient *zclient = NULL;
 /* For registering threads. */
 extern struct thread_master *master;
 
-static struct interface *zebra_interface_if_lookup(struct stream *s)
-{
-	char ifname_tmp[INTERFACE_NAMSIZ];
-
-	/* Read interface name. */
-	stream_get(ifname_tmp, s, INTERFACE_NAMSIZ);
-
-	/* And look it up. */
-	return if_lookup_by_name(ifname_tmp, VRF_DEFAULT);
-}
-
 /* Inteface addition message from zebra. */
 static int sharp_ifp_create(struct interface *ifp)
 {
@@ -102,11 +91,8 @@ static int interface_address_delete(ZAPI_CALLBACK_ARGS)
 	return 0;
 }
 
-static int interface_state_up(ZAPI_CALLBACK_ARGS)
+static int sharp_ifp_up(struct interface *ifp)
 {
-
-	zebra_interface_if_lookup(zclient->ibuf);
-
 	return 0;
 }
 
@@ -385,11 +371,6 @@ static int sharp_nexthop_update(ZAPI_CALLBACK_ARGS)
 	return 0;
 }
 
-static int sharp_ifp_up(struct interface *ifp)
-{
-	return 0;
-}
-
 static int sharp_ifp_down(struct interface *ifp)
 {
 	return 0;
@@ -414,7 +395,6 @@ void sharp_zebra_init(void)
 	zclient_init(zclient, ZEBRA_ROUTE_SHARP, 0, &sharp_privs);
 	zclient->zebra_connected = zebra_connected;
 	zclient->interface_delete = interface_delete;
-	zclient->interface_up = interface_state_up;
 	zclient->interface_down = interface_state_down;
 	zclient->interface_address_add = interface_address_add;
 	zclient->interface_address_delete = interface_address_delete;
