+++ staticd/static_zebra.c
@@ -49,17 +49,6 @@ bool debug;
 struct zclient *zclient;
 static struct hash *static_nht_hash;
 
-static struct interface *zebra_interface_if_lookup(struct stream *s)
-{
-	char ifname_tmp[INTERFACE_NAMSIZ];
-
-	/* Read interface name. */
-	stream_get(ifname_tmp, s, INTERFACE_NAMSIZ);
-
-	/* And look it up. */
-	return if_lookup_by_name(ifname_tmp, VRF_DEFAULT);
-}
-
 /* Inteface addition message from zebra. */
 static int static_ifp_create(struct interface *ifp)
 {
@@ -107,25 +96,18 @@ static int interface_address_delete(ZAPI_CALLBACK_ARGS)
 	return 0;
 }
 
-static int interface_state_up(ZAPI_CALLBACK_ARGS)
+static int static_ifp_up(struct interface *ifp)
 {
-	struct interface *ifp;
+	if (if_is_vrf(ifp)) {
+		struct static_vrf *svrf = static_vrf_lookup_by_id(ifp->vrf_id);
 
-	ifp = zebra_interface_if_lookup(zclient->ibuf);
-
-	if (ifp) {
-		if (if_is_vrf(ifp)) {
-			struct static_vrf *svrf =
-					static_vrf_lookup_by_id(vrf_id);
-
-			static_fixup_vrf_ids(svrf);
-			static_config_install_delayed_routes(svrf);
-		}
-
-		/* Install any static reliant on this interface coming up */
-		static_install_intf_nh(ifp);
-		static_ifindex_update(ifp, true);
+		static_fixup_vrf_ids(svrf);
+		static_config_install_delayed_routes(svrf);
 	}
+
+	/* Install any static reliant on this interface coming up */
+	static_install_intf_nh(ifp);
+	static_ifindex_update(ifp, true);
 
 	return 0;
 }
@@ -499,11 +481,6 @@ extern void static_zebra_route_add(struct route_node *rn,
 			   zclient, &api);
 }
 
-static int static_ifp_up(struct interface *ifp)
-{
-	return 0;
-}
-
 static int static_ifp_down(struct interface *ifp)
 {
 	return 0;
@@ -527,7 +504,6 @@ void static_zebra_init(void)
 	zclient->zebra_capabilities = static_zebra_capabilities;
 	zclient->zebra_connected = zebra_connected;
 	zclient->interface_delete = interface_delete;
-	zclient->interface_up = interface_state_up;
 	zclient->interface_down = interface_state_down;
 	zclient->interface_address_add = interface_address_add;
 	zclient->interface_address_delete = interface_address_delete;
