+++ vrrpd/vrrp_zebra.c
@@ -108,22 +108,9 @@ static int vrrp_zebra_if_del(int command, struct zclient *zclient,
 	return 0;
 }
 
-static int vrrp_zebra_if_state_up(int command, struct zclient *zclient,
-				  zebra_size_t length, vrf_id_t vrf_id)
+int vrrp_ifp_up(struct interface *ifp)
 {
-	struct interface *ifp;
-
-	/*
-	 * zebra api notifies interface up/down events by using the same call
-	 * zebra_interface_state_read below, see comments in lib/zclient.c ifp =
-	 * zebra_interface_state_read(zclient->ibuf, vrf_id);
-	 */
-	ifp = zebra_interface_state_read(zclient->ibuf, vrf_id);
-
-	if (!ifp)
-		return 0;
-
-	vrrp_zebra_debug_if_state(ifp, vrf_id, __func__);
+	vrrp_zebra_debug_if_state(ifp, ifp->vrf_id, __func__);
 
 	vrrp_if_up(ifp);
 
@@ -224,11 +211,6 @@ int vrrp_zclient_send_interface_protodown(struct interface *ifp, bool down)
 						down);
 }
 
-int vrrp_ifp_up(struct interface *ifp)
-{
-	return 0;
-}
-
 int vrrp_ifp_down(struct interface *ifp)
 {
 	return 0;
@@ -250,7 +232,6 @@ void vrrp_zebra_init(void)
 	zclient->zebra_connected = vrrp_zebra_connected;
 	zclient->router_id_update = vrrp_router_id_update_zebra;
 	zclient->interface_delete = vrrp_zebra_if_del;
-	zclient->interface_up = vrrp_zebra_if_state_up;
 	zclient->interface_down = vrrp_zebra_if_state_down;
 	zclient->interface_address_add = vrrp_zebra_if_address_add;
 	zclient->interface_address_delete = vrrp_zebra_if_address_del;
